(PUP-10216) Use last environment if unable to fetch node definition
Previously, when Puppet failed to get its node definition, it would try
to pluginsync against the agent-configured environment (usually
'production'). This is inappropriate, as it first deletes all of its
plugins, and then has to re-sync them fresh from the correct environment
after it switches over.

The environment is now written in the lastrunfile, which Puppet attempts
to read if it fails to fetch the node definition. Environment values set
with `--environment` or in puppet.conf are still prioritized over the
last used environment.