(PUP-6801) Add new setting helpers
Currently Puppet::Settings#set_by_cli? returns true or false, however
it's also useful to know the actual value of the setting looked up by
the method.

Create a #set_by_cli method which returns the value of the setting, and
have #set_by_cli? call it.

Follow the same pattern for #set_by_section?. This method was added in
c641b04b0, but its usage was removed in 79737d7f, so we're safe to
change it.

Document the #set_by_cli and #set_in_section methods and update the
tests.