(maint) Provide types for possible integer settings defaults
The change in eeb61d1c6 allowed settings to be strings containing digits
and to remain strings (helpful for numeric certnames for example).

However, several settings depended on the automatic conversion of
strings that contain digits to Integer. This adds type specification
to any defaults that match / \d+,/ in the defaults.rb.

All are given the type `:integer` unless the setting contains the word
"port" in which case they gain the type `:port`.

A notable exception being the ldapport setting, which depends on being a
string and now is explicitly a string in the defaults.

Additional searches for ttl & priority didn't reveal any defaults that
needed changing. Nor did a search for strings containing wholy digits.