(PUP-11088) Check all limits on top level facts
Puppet doesn't warn the user when the facts exceed certain limits. This commit
adds checks for the number of top level facts and for the lengths of either the
names or values of the said facts. Options for the user to change the values
for these limits were also added.