(maint) Update tests to match more strict settings behavior
Prior to eeb61d1c64, puppet automatically coerced settings whose values were all
digits to an integer. Puppet no longer does that, instead it stores the value
that was set, and converts the value when it is looked up, based on the
settings' type, e.g. :integer, :enum, etc.

The facter_spec change is because `serverport` is a `:port` setting so trying to
set it to "baz" is invalid. It's not a breaking change, because running `puppet
agent -t --serverport baz` would never have worked.

The setting_spec change is because we no longer coerce the string to an integer.