(RFC) Readable type errors when running Bolt plans and tasks
Many Bolt users aren't familiar with Puppet, and type errors are
difficult to read. This proposes only print the short name of types in
type mismatch errors (so "expect Bolt::TargetSpec, got Integer") when
the `Puppet[:tasks]` setting is set. This ensures that this change only
impacts Puppet when running in a Bolt context. The main caveat is that
errors will still be fully detailed when running apply blocks, but
because we expect apply block users to be more advanced and expect many
users to be aware that apply blocks are running in a different context
from the rest of the Bolt plan in general, this is acceptable.