(PUP-11029) Fix pip package provider to query available versions
Starting with version `20.3b1`, `pip` removed available versions of a
package from output when trying to install a wrong or unavailable
version. Puppet is using this output to gather all available versions of
a package and determine the latest one. Even though this regression got
fixed in `21.1` version of `pip`, this commit adds the
`--use-deprecated=legacy-resolver` argument when querying available
versions of the managed package and also using any of the affected
versions of `pip`.