(PUP-11158) Make it possible to guard an environment from eviction
Allow an environment cache entry to be guarded. If guarded, then the environment
won't be reloaded "out from underneath" the caller, e.g. during a compile.

Note the act of guarding or unguarding doesn't cause the environment (or any
other environment) to be evicted.

The guard/unguard methods are added to the base EnvironmentLoader module which
is mixed into various environment loaders, and the methods are overridden in the
Cached loader.

Although multiple puppetserver/JRuby threads can call the environment cache
concurrently, the instance methods are synchronized, so we don't need to lock
the cache entry's counter.