(PUP-11158) Use environment guard to prevent environment from reloading
Previously the `with_text_domain` method was used to add the text domain and
module translations, assuming they would be present for the duration of
compilation. However, we were getting lucky, as resolving per-environment
settings like modulepath, result in a call to `Cached#get_conf(Symbol)` which
would try to clear expired environments. However, the cache keys are Strings,
so the environment wasn't cleared.

This commit adds an environment guard to make the behavior explicit. Within the
block (such as during compilation) the environment will not be reloaded. The
behavior is the same, but doesn't rely on the symbol/string mismatch.