(PUP-10308) Fix Puppet.lookup(:current_environment)
At the start of the configurer run, puppet pushes
its configured environment or the server-specified
environment onto the context.

If pluginsync causes environment to change,
then the agent switches to the new environment but
the new environment is not pushed onto the context,
leaving the old environment set.

This commit updates the `lib/puppet/configurer.rb`
to push the new environment after the convergence
is done.