(PUP-11169) Fix missing trailing `/` for versioned dirs static catalogs
Previous to this change, the `environmentpath` was set using the
internal Puppet::FileSystem.expand_path, which had the side effect
of stripping the value's trailing slash. This broke the regex used
for determining if a file met the criteria for
`inlineable_metadata?` and stopped catalogs from getting any
metadata values for files.

This change moves the trailing slash into the regex, so that the
`environmentpath` is not required to contain the slash to match. It
also ensures that the object returned by `real_path` is a Pathname,
as that is a requirement for the `get_content_uri` method.