(PUP-6802) Skip pluginsync if environment unavailable on server
Previous to this PR, when an agent agent starts a run with an environment not
available on the server, it will also try to pluginsync with the unexistent
environment.

Now, a request is sent to the server checking if there is anything to
pluginsync, or if the environment even exists.