(maint) Remove 'broken' test filtering
Remove the ability to filter `broken` tests, just use `pending` or `skip`.

One test relied on the `broken` filter. It was introduced in 4654ba39, prior to
the future parser. The wacky behavior it mentions no longer occurred after
1a544914 when switching to the future lexer. Now the code:

    class class inherits default {}

Always produces a parse error with the message:

    'class' keyword not allowed at this location

So the "broken" test is removed.