(PUP-11158) Use environment guard to prevent environment from reloading
Previously, per-environment settings like `modulepath` resulted in a call to
`Cached#get_conf(Symbol)` which tried to clear expired environments. However,
environment cache keys are currently Strings, so the environment wasn't cleared.
While that is a bug, it prevented the environment from being reloaded during
compilation.

This commit adds an environment guard to make the behavior explicit. This way we
get the same behavior, but we're not relying on symbol/string mismatch.