(maint) Teardown remaining lastrunfile after using tmp env
Recent work allowed the agent to skip node requests by inspecting a
lastrunfile if it exists and start the agent check-in process with the
environment listed there. However, the acceptance tests often create
dummy enviroments and do not clean up the lastrunfile.

A first pass updated many tests to remove the lastrunfile on all agents
if they used a temporary environment. However, several were missed as
the agent testing will only test a subset of the server tests. In an
attempt to quickly rectify the situation this patch audits all
acceptance tests that use, or mention that they should use,
`mk_tmp_environment_with_teardown` and extends their teardowns to use
the existing pattern for removing the lastrunfile.