(maint) Update expected error message in environment test
As we removed the node request, and the default value of
`ignore_plugin_errors` is now false, we don't get a 404 error from the
server anymore (this could have happened either from the node request or
the catalog request).

To account for this change, update the expected error message to match
the failure.