(maint) Teardown lastrunfile when creating custom env content
These tests have refactor notes to use the
"mk_temp_environment_with_teardown" helper. Which is a typo of the
"mk_tmp_environment_with_teardown" helper. Currently, these tests do not
cause an issue with the lastrunfile because the environment they modify
is production. However, if they were to be refactored to use a custom
environment like they should they would immediately cause regressions,
which may be hard to diagnose.

This adds the lastrunfile teardown step like users of
mk_tmp_environment_with_teardown use to be kinder to our future selves.