(PUP-11170) User => absent deletes homedir on AIX (when managehome=true)
Before this commit, when a manifest that deletes an AIX user is applied,
with homedir set to true, puppet would not delete the home directory for
that user because it was using `rmuser`, and it has no flag to delete
home directories.

In this commit, we switched to a simple folder removal of homedir if
managehome is enabled as userdel also removes the user from
/etc/security/passwd, and rmuser does not. Puppet will not remove home
directory if this is relative, '/', symlink or contains other user home
directories.