(maint) Increase wait time
Tests sometimes failed on Windows because it was possible for the thread to exit
before the agent started up. If the thread didn't exist, then the agent could
forcibly acquire the stale lock and continue the run, failing to connect to a
non-existent server.

This extends the maximum amount of time the worker thread hold onto the lock.