(PUP-11201) Refine when the last used environment is used
With the recent changes in PUP-10216, the agent stores the last used
environment in `last_run_summary.yaml`. If the next run does not specify
an environment on the CLI, then the last used environment always takes
precedence over the environment specified in puppet.conf.

This change makes an exception for the agent-specified case so that the
second run only uses the environment from the previous run if it was
server-specified. Otherwise, the agent should start in its configured
environment (from puppet.conf) or default to "production".