(PUP-10105) puppet resource --to_yaml emits puppet class tags
    Previously some puppet resources would be the objects of
`Puppet::Util::Execution::ProcessOutput` which extends ruby String. When pops
serialization checked whether a resource was a String, it used the method
`is_a?`, which returned True for ProcessOutput since it is a child of String,
resulting in weird yaml output.

    This commit updates the `instance?` methods to use `instance_of?` instead
of `is_a?`, since it only returns True if an object is an object of the exact
class, not a subclass