(PUP-10105) Update the `ToStringifiedConverter#to_data` method to use `instance_of?`
The `to_data` method checked wheter a value was an instance of String with the
is_a? method, which would allow any subclasses of String to pass trough the
logic, failing to properly convert them.

This commit changes the logic to use `instance_of?` when checking Strings, so
any subclasses go trough the `unknown_to_string()` method.