(PUP-10890) Add tests for lookup options
Add tests for empty lookup options and options containing a symbol which is not
a valid Puppet::LookupValue. The tests are added at the environment and module
layers. The empty lookup options at the environment layer do pass, but the
module layer does not, so the latter is marked as pending.