(PUP-10890) Use global lookup options if module options are nil
If lookup_options were defined at the global or environment layer and a module
defined an empty hash of lookup_options, then lookup failed trying to call
`each_pair` on nil.

Now the `validate_lookup_options` method returns early. However, we cannot merge
the nil module options with global options (MergeStrategy#assert_type raises
if the value is nil). So now we only merge if both are non-nil. If one is nil,
but the other is not, we use the other. If both are nil, then we return nil, as
we did before.

Note `compile_patterns` is nil safe.