(PUP-6708) Decouple sending and caching report
Previously, if the agent failed to send its report, then it didn't update its
cached version. So there was no record of the run, aside from syslog messages.
The limitation is due to the way the indirector works. If the main terminus'
save method raises, then the cache terminus is never called.

This commit decouples the main and cache termini, so that we can ensure the
cache is always updated.