(PUP-11188) Do not add auth and cookie header when redirecting
When following HTTP redirects the authentication and cookie header is passed to
different hosts, which may leak sensitive information.

Now the authentication and cookie header is not added when redirecting to
different hosts.  The headers are added when redirecting to the same host, or
when adding the new Puppet setting `:location_trusted`.