(PUP-11209) Puppet agent silently skips unknown resources.
Previously all unknown resources were converted into a component
(Puppet::Type::Component) and silently skipped.

This commit adds a new resource attribute (:kind) that specifies the resource
kind, which is used to differentiate between built-in types and user defined
types.

When converting the catalog to a RAL catalog, this new attribute is used to to
check whether a resource should use the resource type's :new method, or create
a component, and in the case where the server considers a resource to be
compilable, but the agent can't resolve it, then we raise an error