(PUP-11237) Detect recursive json generation
If the `MultiJson` gem was absent, then `Puppet::Util::Json.dump` failed if
given an array or hash containing an indirected object. The issue occurred because the
json library will call the object's `to_json` method (since it's not a primitive
type) passing an instance of `JSON::State` as the `options` argument. The
`to_json` method for indirected objects is implemented in
`Puppet::Network::FormatSupport`, and that calls `Puppet::Util::Json.dump`
recursively.

Now we detect if the `options` is a `JSON::State` object, similar to how the
JSON library detects if it's being called recursively[1].

https://github.com/ruby/ruby/blob/v2_7_3/ext/json/generator/generator.c#L1071-L1088