(PUP-11217) do not eager load http_client
https://github.com/puppetlabs/puppet/pull/8749/files#diff-91bcbee7bf547ed6ab2546124a1a87c5e34d8189ef781252293b65725ea40bddR145 introduced a change
on how we load runtime services, setting them
upfront before running any tests. This breaks the
ability to mock the implementations, and enforces
specific implementations to be injected when Puppet
is initialized. To avoid breaking other projects,
do not load the HttpClient for now, but let it be
lazy initailized as it was before #8749.