(maint) Regenerate el-repo fixtures
If FIPS is enabled on RedHat 8, RPM validation becomes more secure.

If a vendor supplies an RPM that is unsigned or weakly-signed or has
digests that use use weak algorithms, yum/dnf will refuse to install the
package. This seemed to be the case with the existing el-repo fixtures:

```
> yum install kernel-devel-puppet
...
Error: Transaction check error:
  package kernel-devel-puppet-3.10.0-1062.4.3.noarch does not verify: no digest

> rpm -v --checksig /tmp/el-repo/RPMS/kernel-devel-puppet-3.10.0-1062.1.1.noarch.rpm
/tmp/el-repo/RPMS/kernel-devel-puppet-3.10.0-1062.1.1.noarch.rpm:
    Header SHA1 digest: OK
    Payload SHA256 digest: NOTFOUND
    MD5 digest: NOTFOUND
```

This can be worked around by passing `--nodigest` and `--nofiledigest`,
but this only works on the rpm command, not yum (which is used by
puppet), so the only way to make this test work is by rebuilding the
packages with the correct digests.