(PUP-11265) Push last used environment and loaders onto the context
Previously, if the agent run decided to use the last server environment,
it would not push the current environment and loaders onto the context.

Metadata requests that don't specify an environment will cause the
indirector to fall back to the environment stored in the context[1],
which in our case is the environment the agent started in[2] instead of
the converged one.

Push the current environment and loaders if we're using the last server
environment, and extract the logic to a separate method.

[1] https://github.com/puppetlabs/puppet/blob/3828aabe8d32368faa5cdc1a428189d0cb117e52/lib/puppet/indirector/request.rb#L33
[2] https://github.com/puppetlabs/puppet/blob/3828aabe8d32368faa5cdc1a428189d0cb117e52/lib/puppet/application/agent.rb#L389-L401