(PUP-11315) Allow resource attributes named "plan" and "apply"
Previously, using either of these keywords as resource attribute names
within an apply() block in a plan would fail. That happened because they
weren't included in the `keyword` rule. This was only a problem when
running in a plan, because those keywords only exist when the `tasks`
setting is set. In normal Puppet operation, they aren't lexed as special
tokens at all, so they didn't need to be in the `keyword` rule to
function properly.

We now add them to the `keyword` rule regardless of whether they are
actually lexed as keywords, which works fine in both cases.