(PUP-7362) Fail fast if there are no facts for a node
Previously, puppet lookup would return incorrect results for any target
node that had no facts cached.

For example when using the `Per-node data` hierarchy, with the path to
the data fail being the value of `$trusted[certname]`, a node that had
the said fact cached would return the correct data from its data file,
but a node which didn't have any facts cached, would default to a common
configuration file, outputting the data from there.

This is misleading since it doesn't error out, giving the impression
that the run went well, when in actuality it failed to fetch the correct
data.

This commit fixes this issue by failing the run the moment we are able
to see that the target node has no cached facts.