(PUP-10435) Facts provided in a file can't be used for classification
5cde767dc1e67afc729b4ff6cdf5d603e02228ae added the ability to use facts
from a fact file, but those facts were added after the classification
happened, meaning that any facts used as rules for classification
couldn't be overridden with the facts from the fact file.

This commit fixes this issue by querying pdb to get the target node's
facts, and merge the facts provided from the fact file, before the
classification happens.