(PUP-11236) Work around Thread#thread_variable? bug
Thread#thread_variable? sometimes returns false[1] on MRI, which causes the
ThreadLocalSingleton to not behave like a singleton. Instead call `thread_variable_get`
to see it's set. This means the singleton value can't be `nil`, but we
never do that in practice.

Kudos to @mikaelsmith for finding the ruby bug!

[1] https://bugs.ruby-lang.org/issues/16906