(PUP-11322) Append parse results
Each time we parsed a manifest, we copied the already merged code for the `main`
hostclass to a new BlockExpression and then added the code for the one new
parsed result. This process repeated for each parsed manifest leading to N^2
memory allocation where N is the number of manifests.

This creates a new method `append` that should be used to append a new parsed
result to an existing set of results. In order to preserve the flattening
behavior, if the `left` results haven't been flattened yet, then we fallback
to concatenate.