(maint) Fix last server environment test
In bb1017f9 we changed the agent's behavior to fall back to making a
node request if the lastrunfile does not contain a usable environment.

Update the `use_last_server_specified_environment` test to reflect this
change in behavior. Based on test ordering, this test could have passed
if the `lastrunfile` was present when the first agent run started. To
ensure the new behavior is triggered, always clean up the `lastrunfile`
at the beginning of the test.