(PUP-5800) Update to more recent Tempfile
Update our Uniquefile to use a more recent Tempfile implementation from 2.7.3
while preserving how we generate the unique filename, locking behavior and
defaulting to mode 0600.

One different behavior is Uniquefile will expand tmpdirs prior to checking if it
is a directory, writable, etc, and an error is raised if no tmpdir is found.

One test had to be updated because it relied on passing '/' in the basename, but
Uniquefile strips those from the basename. So instead pass the non-existent lock
as the `tmpdir`.