(PUP-5800) Use Dir::Tmpname to generate the unique name
The `Tempfile` functionality for generating a unique name was extracted into the
`Dir::Tmpname` module, so use that. It retries in the same way that we used to,
but is more strict about deleting unusuable chars[1], [2], and it no longer
calls File.expand_path as that can expand '~' in the basename[3], which was the
cause of this ticket.

[1] https://github.com/ruby/ruby/commit/59cfa0b1346067e36ae11fc6c317693ce4b9833a
[2] https://github.com/ruby/ruby/commit/10b96900b90914b0cc1dba36f9736c038db2859d
[3] https://github.com/ruby/ruby/commit/6f8bce9eff15b527c87e1b13312db2fbf0b504f4