(PUP-11320) Move `ssh_authorized_key` resources creation at the end
This commit moves the moment when `ssh_authorized_key` resources are
created from the beginning of the user type evaluation, to the end of it
to avoid order dependency issues.

PUP-11067 fixed the case where a user gets created and at the same time
any ssh keys associated are removed in the same manifest. To achieve
this, the ensure property gained priority over generating the ssh key
removal resources as these depend on the user home location.

PUP-11320 allows once again natural priority of the user properties and
parameters but moves the `ssh_authorized_key` resources creation at the
end of the user type flow, after all user properties and parameters were
resolved.