(PUP-11321) Allow loading of safe classes from the persistence store
The result of a deferred function is stored in the persistence store as the last
known "system_value". When calling the `binary_file` function deferred, we
stored a `Puppet::Pops::Types::PBinaryType::Binary` object, but failed to load
it the next time we ran.

Expand the list of allowed classes to include objects that may be called by
builtin or custom functions.