(PUP-11320) Ignore authorized key purging if home doesn't exist yet
It wasn't possible to create a user and specifiy "purge_ssh_keys => true" or an
explicit path because the user type relies on "generate" to create
"ssh_authorized_keys" resources with "ensure" set to "absent". Ideally we would
use "eval_generate" to ensure the user is created and its home directory exists
before resolving the path to authorized keys. However, that's blocked on
PUP-2718.

This commit instead ignores the purge logic if the user's home dir doesn't exist
and "purge_ssh_keys" depends on the home dir to resolve the authorized_key path.
That happens in the case of purge_ssh_keys=true or set to a string or array,
where the entry starts with ~/ or %h/.