(maint) Do not output raw serialization in acceptance
Previously, we would output the full content of an agent run using
http_debug when checking serialization formats. We require checking
the stdout of the http_debug command to validate the content-type
header was set correctly.

However, outputting the raw bytes from the http requests to the logs
can be problematic because in CI we will take each test's output and
place them into xml files for Jenkins to display. These raw bytes may
contain invalid xml content that causes the post-run CI actions to fail
the build.

This sets the command which uses http_debug to "silent" so that its
content will not be sent to stdout.