(PUP-11433) Use systemd by default in CentOS/RHEL
Since CentOS7/RHEL7, systemd is used to manage services. This change
changes the default provider in CentOS/RHEL from the old redhat
provider based on init, to the systemd provider, so that we no longer
need to update the provider code when adding support for a new version
of CentOS/RHEL.

The old redhat provider is still used for CentOS/RHEL<7 which are using
init to manage services.