(PUP-11456) Avoid ruby kwarg and hash confusion with doubles
Previously the test failed on ruby 3.0.3 and rspec 3.10.3 due to expecting a
method to be called with keyword arguments, but getting called with a hash.
The issue is described in [1], but the problem stems from using a "double" since
rspec has no way of knowing whether the "real" object accepts keyword args or a
hash. This commit avoids the issue entirely by using a real object.

[1] https://github.com/rspec/rspec-mocks/issues/1460