(PUP-11459) Always load Puppet::Util::Windows
Previously, 'puppet/util/windows.rb' was loaded on Windows due to requiring
'puppet/file_system/uniquefile'[1]. The second require[2] was actually a noop,
since the file had already been required. And on non-Windows, the file was never
loaded.

Now we explicitly require 'puppet/util/windows' on all platforms. It must be
done after 'puppet/util/platform' so we can detect if we're actually on Windows.
We also remove the misleading requires that didn't do anything.

[1] https://github.com/puppetlabs/puppet/blob/6.26.0/lib/puppet/util.rb#L11
[2] https://github.com/puppetlabs/puppet/blob/6.26.0/lib/puppet/util.rb#L25