(PUP-11440) If no env found, and strict env mode cancel puppet run
Previously if you were running the puppet agent with strict enviroment mode
and the given enviroment was not found, puppet would then default to the production
environment.
This can produce unexpected results, so now we are going to fail early when
the environment is not found and strict environment mode is on.