(PUP-11471) Make using default path an option of create_x509_store
OpenSSL::X509::Store are created by #create_x509_store. The store is
often kept as is but in the context of #create_system_context it is
modified to also include system certificates.

In order to add support for system certificates to more SSL Contexts,
add a parameter to include or not the system's certificates.

No functional change.