(PUP-9001) Show diff for new files
Prior to this commit when :show_diff is set, file diffs are only shown
when modifying an existing file. However, it is often very helpful to
see the diff for a new file as well, either in testing or production
logs.

After this commit diffs are shown for new files as well. Diffs are
performed against NUL(windows) or (/dev/null)(all others) to ensure
compatibility with custom diff commands.