(PUP-11428) Reload ssl state for each agent run
Previously, we only loaded the agent's ssl state once when the agent started,
but not between runs. Since the agent may sleep for upwards of `runinterval`
seconds and those files may change on disk, reload them each time.

The `wait_for_certificates` behavior depends on whether we're running onetime or
not and whether `waitforcert` was specified on the `puppet agent` command line.
So pass those options through.

Enable pending test now that it passes.