(PUP-11428) Add integration tests
Adds 2 tests for the agent ssl bootstrapping. Also adds pending test for
reloading a CRL between runs. It performs an agent run which overwrites the
agent's cert and key (to one that is revoked). The second run should then fail,
but doesn't currently. We stub the `run_event_loop` to ensure we don't
get into an infinite loop.