(PUP-11454) Enforce salt requirements for macOS versions 10.15+
For macOS versions 10.15 and higher, it is required that the salt is 32 bytes.
user.rb was modified so the salt requirements were included.
directoryservice.rb was modified so the salt requirements are enforced.
should_allow_managed_macos_users_to_login.rb was modified so it would test on macOS versions 10 and higher, not just 11.