(PUP-9323) Resolve deferred values at resource evaluation time
Allow deferred values to be resolved lazily as each resource is evaluated.
This also means validation, munging and unmunging are delayed until the
resolved value is set back on the parameter or property.

The deferred evaluation is similar to how Puppet::Type#eval_generate works, but
since we're not modifying the graph, we don't need to worry about containment
edges, tags, noop, etc.